requirements.txt
pandas==1.5.3
scipy==1.10.1
numpy==1.24.2
uvicorn==0.20.0
fastapi==0.95.0
scikit-learn==1.2
tables==3.8.0
pyarrow==11.0.0
feather-format==0.4.1
geopy==2.3.0
requests==2.28.2
httpx==0.23.3
pyjwt==2.6.0
python-multipart==0.0.6
pytest==7.2.2
pytest-asyncio==0.21.0
lightfm==1.17
matplotlib==3.7.1
jinja2==3.1.2